+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.5
+
+- refactor tooltip
+- add screen scroll layer
+
 ## 0.0.2
 
 - Optimize